fix: initialize reverse proxy forwarder with right public certs
public certs were missing for reverse proxy forwarder
due to the fact that it was initialized very early on
even before certs were loaded, move the code such that
we only initialize the forwarder after TLS has been
successfully loaded.
